+++ x10/core/amount.py
@@ -2,7 +2,7 @@ import decimal
 from dataclasses import dataclass
 from decimal import Decimal
 
-from x10.perpetual.assets import Asset
+from x10.models.asset import Asset
 
 ROUNDING_SELL_CONTEXT = decimal.Context(rounding=decimal.ROUND_DOWN)
 ROUNDING_BUY_CONTEXT = decimal.Context(rounding=decimal.ROUND_UP)
@@ -46,6 +46,7 @@ class StarkAmount:
         return StarkAmount(-self.value, self.asset)
 
 
+# FIXME: Remove
 # @dataclass
 # class StarkOrderAmounts:
 #     collateral_amount_internal: HumanReadableAmount
