+++ x10/models/client.py
@@ -1,4 +1,4 @@
-from x10.utils.model import X10BaseModel
+from x10.models.base import X10BaseModel
 
 
 class ClientModel(X10BaseModel):
