+++ x10/models/withdrawal.py
@@ -1,27 +1,27 @@
 from decimal import Decimal
 
-from x10.utils.model import HexValue, SettlementSignatureModel, X10BaseModel
+from x10.models.base import HexValue, SettlementSignatureModel, X10BaseModel
 
 
-class Timestamp(X10BaseModel):
+class TimestampModel(X10BaseModel):
     seconds: int
 
 
-class StarkWithdrawalSettlement(X10BaseModel):
+class StarkWithdrawalSettlementModel(X10BaseModel):
     recipient: HexValue
     position_id: int
     collateral_id: HexValue
     amount: int
-    expiration: Timestamp
+    expiration: TimestampModel
     salt: int
     signature: SettlementSignatureModel
 
 
-class WithdrawalRequest(X10BaseModel):
+class WithdrawalRequestModel(X10BaseModel):
     account_id: int
     amount: Decimal
     description: str | None
-    settlement: StarkWithdrawalSettlement
+    settlement: StarkWithdrawalSettlementModel
     chain_id: str
     quote_id: str | None = None
     asset: str
