+++ x10/perpetual/simple_client/__init__.py
@@ -1,0 +1,3 @@
+from x10.perpetual.stream_client.stream_client import (  # noqa: F401
+    PerpetualStreamClient,
+)
